fix: change Nervape#classId form `uint16` to `uint256`

+++ contracts/INervape.sol
@@ -6,13 +6,13 @@ import "@openzeppelin/contracts-upgradeable/token/ERC721/extensions/IERC721Enume
 interface INervape is IERC721EnumerableUpgradeable {
     function bridgeMint(address to, uint256 tokenId) external;
 
-    function mint(uint16 classId, address to) external returns (uint256);
+    function mint(uint256 classId, address to) external returns (uint256);
 
-    function totalSupplyOfClass(uint16 classId) external view returns (uint16);
+    function totalSupplyOfClass(uint256 classId) external view returns (uint256);
 
-    function maxSupplyOfClass(uint16 classId) external view returns (uint16);
+    function maxSupplyOfClass(uint256 classId) external view returns (uint256);
 
-    function classOf(uint256 tokenId) external view returns (uint16);
+    function classOf(uint256 tokenId) external view returns (uint256);
 
-    function mintable(uint16 classId) external view returns (uint16);
+    function mintable(uint256 classId) external view returns (uint256);
 }
